+++ composeApp/src/desktopMain/kotlin/com/clipevery/clip/service/HtmlItemService.kt
@@ -32,7 +32,7 @@ class HtmlItemService(appInfo: AppInfo) : ClipItemService(appInfo) {
     }
 
     override fun createPreClipItem(
-        clipId: Int,
+        clipId: Long,
         itemIndex: Int,
         identifier: String,
         transferable: Transferable,
@@ -46,7 +46,7 @@ class HtmlItemService(appInfo: AppInfo) : ClipItemService(appInfo) {
     }
 
     override fun doLoadRepresentation(transferData: Any,
-                                      clipId: Int,
+                                      clipId: Long,
                                       itemIndex: Int,
                                       dataFlavor: DataFlavor,
                                       dataFlavorMap: Map<String, List<DataFlavor>>,
